add attendees data, rename to pushlogv2
