+++ src/commands/pushlogv2.ts
@@ -1,11 +1,11 @@
 import { ApplicationCommandData, CommandInteraction } from "discord.js";
 
-import { SessionLog } from "../sessionLog";
+import { SessionEvent, SessionLog } from "../sessionLog";
 import { PushLogData } from "../pushlogTarget";
 import { DateTime } from "luxon";
 
 export const signature: ApplicationCommandData = {
-	name: "pushlogtarget",
+	name: "pushlogv2",
 	description: "Pushes the specified session's logs to an external archive",
 	options: [
 
@@ -18,7 +18,7 @@ export const signature: ApplicationCommandData = {
 
 		{
 			name: "mentors",
-			description: "Mentor Discord ID(s) (e.g.: \"@mentor1,@mentor2\")",
+			description: "Mentor Discord ID(s) (e.g.: \"@mentor1 @mentor2\")",
 			type: "STRING",
 			required: true
 		},
@@ -60,7 +60,6 @@ export async function exec(interaction: CommandInteraction) {
 	const data = toPushData(logToPush, argv.getString("topic-id")!, argv.getString("documentator")!, argv.getString("mentors")!);
 
 	const pushResult = await global.pushlogTarget.push(data);
-
 	await interaction.editReply(`>>> Attempted to push to archive. Result: ${pushResult}`);
 }
 
@@ -71,6 +70,28 @@ function toPushData(sessionLog: SessionLog, topicId: string, recorderName: strin
 		sessionTimeISO: sessionLog.timeStarted.toUTC().toISOTime(),
 		durationISO: DateTime.fromMillis(sessionLog.timeEnded.toMillis() - sessionLog.timeStarted.toMillis()).toUTC().toISOTime(),
 		recorderName: recorderName,
-		mentorDiscordUserIds: mentorDiscordUserIdsInput.split(",").map(id => id.replace("<@!", "").replace(">", ""))
+		mentorDiscordUserIds: mentorDiscordUserIdsInput.split(" ").map(id => id.replace("<@!", "").replace(">", "")),
+		attendees: Array.from(arrayGroupBy(sessionLog.events, (event) => event.userId).entries()).map(([userId, events]) => {
+			return {
+				discordUserId: userId,
+				attendanceDurationISO: DateTime.fromMillis(
+					events.reduce(
+						(duration, event) => duration + ((event.timeOccurred.toMillis() - sessionLog.timeStarted.toMillis()) * (event.type === "Join" ? -1 : 1)), 0))
+						.toUTC()
+						.toISOTime()
+			}
+		})
 	}
+}
+
+function arrayGroupBy<T>(array: Array<T>, grouper: (value: T) => string): Map<string, Array<T>> {
+	const result = new Map<string, Array<T>>();
+	for (const value of array) {
+		const key = grouper(value);
+		if (!result.has(key)) {
+			result.set(key, []);
+		}
+		result.get(key)!.push(value);
+	}
+	return result;
 }
